Cross Examination Animation
- Added the cross examination animation
- Fixed a bug where the WT and CE animation couldn't be executed more than once
- Cleanup parseScene()
- main.c: switched from malloc to calloc to prevent the use of unitialized values

+++ Engine/include/Menus/Dialogue.h
@@ -34,7 +34,7 @@
     } DialogueContext;
 
     void ClearDialogueText(DialogueContext* Context);
-    int SetDialogueText(DialogueContext* Context, char* Name, char* Text, char SndEffType);
+    size_t SetDialogueText(DialogueContext* Context, char* Name, char* Text, char SndEffType);
     DialogueContext* InitDialog(DisplayDevice* DDevice, BitmapFont* MainFont, BitmapFont* NameFont);
     void Dialogue(DialogueContext* Context);
